@@ -4,68 +4,68 @@
   "private": true,
   "type": "commonjs",
   "scripts": {
-    "postinstall": "husky install || true",
+    "postinstall": "husky || true",
     "dev": "prisma migrate dev && nest start --watch",
-    "build": "node -e \"fs.promises.rm('dist', { recursive: true, force: true })\" && prisma generate && nest build",
-    "start": "prisma migrate deploy && NODE_ENV=production node dist/server/main.js",
-    "format": "prisma generate && prettier --write src test",
+    "build": "node -e \"fs.promises.rm('.build', { recursive: true, force: true })\" && prisma generate && nest build",
+    "start": "prisma migrate deploy && NODE_ENV=production node .build/server/main.js",
     "lint": "prisma generate && tsc --noEmit && eslint --fix src test",
-    "test:e2e": "prisma generate && vitest run --config ./vitest.config.e2e.js",
+    "test:e2e": "prisma generate && vitest run --config ./vitest.config.e2e.mjs",
     "test": "prisma generate && vitest run"
   },
   "dependencies": {
-    "@apollo/server": "^4.9.3",
-    "@as-integrations/fastify": "^2.1.0",
-    "@fastify/static": "^6.11.2",
-    "@nestjs/apollo": "^12.0.9",
-    "@nestjs/common": "^10.2.6",
-    "@nestjs/core": "^10.2.6",
-    "@nestjs/graphql": "^12.0.9",
-    "@nestjs/jwt": "^10.1.1",
-    "@nestjs/mapped-types": "^2.0.2",
-    "@nestjs/passport": "^10.0.2",
-    "@nestjs/platform-fastify": "^10.2.6",
-    "@prisma/client": "^5.3.1",
-    "class-transformer": "^0.5.1",
-    "class-validator": "^0.14.0",
-    "dataloader": "^2.2.2",
-    "fastify": "^4.23.2",
-    "graphql": "^16.8.1",
-    "husky": "^8.0.3",
-    "passport": "^0.6.0",
-    "passport-jwt": "^4.0.1",
-    "prisma": "^5.3.1",
-    "reflect-metadata": "^0.1.13",
-    "rxjs": "^7.8.1",
-    "source-map-support": "^0.5.21"
+    "@apollo/server": "4.11.2",
+    "@as-integrations/fastify": "2.1.1",
+    "@fastify/static": "7.0.4",
+    "@nestjs/apollo": "12.2.1",
+    "@nestjs/common": "10.4.6",
+    "@nestjs/core": "10.4.6",
+    "@nestjs/graphql": "12.2.1",
+    "@nestjs/jwt": "10.2.0",
+    "@nestjs/mapped-types": "2.0.5",
+    "@nestjs/platform-fastify": "10.4.6",
+    "@prisma/client": "5.21.1",
+    "class-transformer": "0.5.1",
+    "class-validator": "0.14.1",
+    "dataloader": "2.2.2",
+    "drizzle-orm": "0.36.0",
+    "fastify": "4.28.1",
+    "graphql": "16.9.0",
+    "graphql-subscriptions": "2.0.0",
+    "graphql-ws": "5.16.0",
+    "husky": "9.1.6",
+    "prisma": "5.21.1",
+    "reflect-metadata": "0.2.2",
+    "rxjs": "7.8.1",
+    "source-map-support": "0.5.21"
   },
   "devDependencies": {
-    "@nestjs/cli": "^10.1.18",
-    "@nestjs/schematics": "^10.0.2",
-    "@nestjs/testing": "^10.2.6",
-    "@swc/cli": "^0.1.62",
-    "@swc/core": "^1.3.91",
-    "@types/node": "^20.8.2",
-    "@types/passport": "^1",
-    "@types/passport-jwt": "^3.0.10",
-    "@types/supertest": "^2.0.14",
-    "@typescript-eslint/eslint-plugin": "^6.7.4",
-    "@typescript-eslint/parser": "^6.7.4",
-    "eslint": "^8.50.0",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-airbnb-typescript": "^17.1.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-import": "^2.28.1",
-    "eslint-plugin-prettier": "^5.0.0",
-    "eslint-plugin-react": "^7.33.2",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-simple-import-sort": "^10.0.0",
-    "eslint-plugin-unused-imports": "^3.0.0",
-    "prettier": "^3.0.3",
-    "supertest": "^6.3.3",
-    "typescript": "^5.2.2",
-    "unplugin-swc": "^1.4.2",
-    "vitest": "^0.34.6"
+    "@eslint/compat": "1.2.2",
+    "@eslint/js": "9.14.0",
+    "@nestjs/cli": "10.4.5",
+    "@nestjs/schematics": "10.2.3",
+    "@nestjs/testing": "10.4.6",
+    "@swc/cli": "0.4.0",
+    "@swc/core": "1.7.42",
+    "@types/eslint__js": "8.42.3",
+    "@types/node": "22.8.7",
+    "@types/supertest": "6.0.2",
+    "@typescript-eslint/eslint-plugin": "8.12.2",
+    "@typescript-eslint/parser": "8.12.2",
+    "better-typescript-lib": "2.9.0",
+    "drizzle-prisma-generator": "0.1.7",
+    "eslint": "9.14.0",
+    "eslint-config-prettier": "9.1.0",
+    "eslint-import-resolver-typescript": "3.6.3",
+    "eslint-plugin-import": "2.31.0",
+    "eslint-plugin-simple-import-sort": "12.1.1",
+    "eslint-plugin-unused-imports": "4.1.4",
+    "globals": "15.11.0",
+    "prettier": "3.3.3",
+    "supertest": "7.0.0",
+    "typescript": "5.6.3",
+    "typescript-eslint": "8.12.2",
+    "unplugin-swc": "1.5.1",
+    "vitest": "2.1.4"
   },
-  "packageManager": "yarn@3.6.1"
+  "packageManager": "yarn@4.5.1"
 }
