@@ -7,7 +7,7 @@ export class AppResolver {
   constructor(private readonly appService: AppService) {}
 
   @Query(() => String)
-  getHello(): string {
-    return this.appService.getHello();
+  health(): string {
+    return this.appService.getHealth();
   }
 }
