@@ -4,41 +4,41 @@ import { User } from '@prisma/client';
 
 import { UserModel } from '@/users/models/user.model';
 
+import { BearerAccess } from './decorators/bearer-access';
 import { CurrentUser } from './decorators/current-user';
-import { JwtAccess } from './decorators/jwt-access';
 import { LoginInput } from './dto/login.input';
 import { RefreshTokenInput } from './dto/refresh-token.input';
 import { RegisterInput } from './dto/register.input';
-import { AuthModel } from './models/auth.model';
-import { AuthService } from './auth.service';
+import { TokenModel } from './models/token.model';
+import { AuthService } from './services/auth.service';
 
-@Resolver()
+@Resolver(() => TokenModel)
 export class AuthResolver {
   constructor(private readonly authService: AuthService) {}
 
-  @JwtAccess()
+  @BearerAccess()
   @Query(() => UserModel)
-  async getCurrentUser(@CurrentUser() user: User): Promise<UserModel> {
+  currentUser(@CurrentUser() user: User): UserModel {
     return user;
   }
 
-  @Mutation(() => AuthModel)
-  async login(@Args('body', { type: () => LoginInput }) body: LoginInput): Promise<AuthModel> {
-    const user = await this.authService.loginByCredentials(body);
+  @Mutation(() => TokenModel)
+  async login(@Args('input', { type: () => LoginInput }) input: LoginInput): Promise<TokenModel> {
+    const user = await this.authService.loginByCredentials(input);
 
     return this.authService.login(user);
   }
 
-  @Mutation(() => AuthModel)
-  async register(@Args('body', { type: () => RegisterInput }) body: RegisterInput): Promise<AuthModel> {
-    const user = await this.authService.registerByCredentials(body);
+  @Mutation(() => TokenModel)
+  async register(@Args('input', { type: () => RegisterInput }) input: RegisterInput): Promise<TokenModel> {
+    const user = await this.authService.registerByCredentials(input);
 
     return this.authService.login(user);
   }
 
-  @Mutation(() => AuthModel)
-  async refreshToken(@Args('body', { type: () => RefreshTokenInput }) body: RefreshTokenInput): Promise<AuthModel> {
-    const user = await this.authService.refreshTokenByCredentials(body);
+  @Mutation(() => TokenModel)
+  async refreshToken(@Args('input', { type: () => RefreshTokenInput }) input: RefreshTokenInput): Promise<TokenModel> {
+    const user = await this.authService.refreshTokenByCredentials(input);
 
     return this.authService.login(user);
   }
