@@ -2,6 +2,6 @@ import { Field, InputType } from '@nestjs/graphql';
 
 @InputType()
 export class RefreshTokenInput {
-  @Field(() => String)
+  @Field()
   refreshToken!: string;
 }
