@@ -6,10 +6,10 @@ import { User } from '@prisma/client';
 @InputType()
 export class RegisterInput implements Pick<User, 'email' | 'password'> {
   @IsEmail()
-  @Field(() => String)
+  @Field()
   email!: string;
 
   @MinLength(8)
-  @Field(() => String)
+  @Field()
   password!: string;
 }
