@@ -4,7 +4,7 @@ import { JwtService } from '@nestjs/jwt';
 import { Prisma, User } from '@prisma/client';
 
 import { CryptoService } from '@/common/services/crypto.service';
-import { UsersService } from '@/users/users.service';
+import { UsersService } from '@/users/services/users.service';
 
 @Injectable()
 export class AuthService {
@@ -17,11 +17,15 @@ export class AuthService {
   async loginByJwt(payload: { id: number }, accessToken: string) {
     const user = await this.usersService.findUserById(payload.id);
 
-    if (!user || !user.accessToken) throw new UnauthorizedException();
+    if (!user.accessToken) {
+      throw new UnauthorizedException();
+    }
 
     const match = this.cryptoService.compare(accessToken, user.accessToken);
 
-    if (!match) throw new UnauthorizedException();
+    if (!match) {
+      throw new UnauthorizedException();
+    }
 
     return user;
   }
@@ -37,20 +41,24 @@ export class AuthService {
   async refreshTokenByCredentials(input: { refreshToken: string }) {
     let id: number;
     try {
-      ({ id } = this.jwtService.verify(input.refreshToken, {
+      ({ id } = await this.jwtService.verifyAsync(input.refreshToken, {
         secret: process.env.REFRESH_TOKEN_SECRET,
       }));
-    } catch (err) {
+    } catch {
       throw new UnauthorizedException();
     }
 
     const user = await this.usersService.findUserById(id);
 
-    if (!user || !user.refreshToken) throw new UnauthorizedException();
+    if (!user.refreshToken) {
+      throw new UnauthorizedException();
+    }
 
     const match = this.cryptoService.compare(input.refreshToken, user.refreshToken);
 
-    if (!match) throw new UnauthorizedException();
+    if (!match) {
+      throw new UnauthorizedException();
+    }
 
     return user;
   }
@@ -60,15 +68,16 @@ export class AuthService {
       id: user.id,
     };
 
-    const accessToken = this.jwtService.sign(payload, {
-      secret: process.env.ACCESS_TOKEN_SECRET,
-      expiresIn: process.env.ACCESS_TOKEN_EXPIRES_IN,
-    });
-
-    const refreshToken = this.jwtService.sign(payload, {
-      secret: process.env.REFRESH_TOKEN_SECRET,
-      expiresIn: process.env.REFRESH_TOKEN_EXPIRES_IN,
-    });
+    const [accessToken, refreshToken] = await Promise.all([
+      this.jwtService.signAsync(payload, {
+        secret: process.env.ACCESS_TOKEN_SECRET,
+        expiresIn: process.env.ACCESS_TOKEN_EXPIRES_IN,
+      }),
+      this.jwtService.signAsync(payload, {
+        secret: process.env.REFRESH_TOKEN_SECRET,
+        expiresIn: process.env.REFRESH_TOKEN_EXPIRES_IN,
+      }),
+    ]);
 
     await this.usersService.updateUser(user.id, {
       accessToken: this.cryptoService.hash(accessToken),
