@@ -1,53 +1,59 @@
 import { Args, Int, Mutation, Query, Resolver } from '@nestjs/graphql';
 
-import { JwtAccess, Role } from '@/auth/decorators/jwt-access';
+import { BearerAccess, Role } from '../auth/decorators/bearer-access';
 
 import { CreateTranslationInput } from './dto/create-translation.input';
 import { UpdateTranslationInput } from './dto/update-translation.input';
 import { UpsertTranslationInput } from './dto/upsert-translation.input';
 import { TranslationModel } from './models/translation.model';
-import { TranslationsService } from './translations.service';
+import { TranslationsService } from './services/translations.service';
 
 @Resolver()
 export class TranslationsResolver {
   constructor(private readonly translationsService: TranslationsService) {}
 
   @Query(() => [String])
-  async getAvailableLanguages(): Promise<string[]> {
+  async languages(): Promise<string[]> {
     return this.translationsService.getAvailableLanguages();
   }
 
   @Query(() => [TranslationModel])
-  async getTranslationsByLanguage(@Args('language') language: string): Promise<TranslationModel[]> {
+  async translations(@Args('language') language: string): Promise<TranslationModel[]> {
     return this.translationsService.findManyByLanguage(language);
   }
 
-  @JwtAccess([Role.Admin])
+  @BearerAccess([Role.Admin])
   @Mutation(() => Boolean)
   async upsertTranslations(
-    @Args('body', { type: () => [UpsertTranslationInput] }) body: UpsertTranslationInput[],
+    @Args('inputs', { type: () => [UpsertTranslationInput] }) inputs: UpsertTranslationInput[],
   ): Promise<boolean> {
-    return this.translationsService.upsertTranslations(body);
+    return this.translationsService.upsertTranslations(inputs);
   }
 
-  @JwtAccess([Role.Admin])
+  @BearerAccess([Role.Admin])
   @Mutation(() => TranslationModel)
   async createTranslation(
-    @Args('body', { type: () => CreateTranslationInput }) body: CreateTranslationInput,
+    @Args('input', { type: () => CreateTranslationInput }) input: CreateTranslationInput,
   ): Promise<TranslationModel> {
-    return this.translationsService.createTranslation(body);
+    const translation = await this.translationsService.createTranslation(input);
+
+    await this.translationsService.pubSub.publish('translationAdded', {
+      translationAdded: translation.createdAt,
+    });
+
+    return translation;
   }
 
-  @JwtAccess([Role.Admin])
+  @BearerAccess([Role.Admin])
   @Mutation(() => TranslationModel)
   async updateTranslation(
     @Args('id', { type: () => Int }) id: number,
-    @Args('body', { type: () => UpdateTranslationInput }) body: UpdateTranslationInput,
+    @Args('input', { type: () => UpdateTranslationInput }) input: UpdateTranslationInput,
   ): Promise<TranslationModel> {
-    return this.translationsService.updateTranslation(id, body);
+    return this.translationsService.updateTranslation(id, input);
   }
 
-  @JwtAccess([Role.Admin])
+  @BearerAccess([Role.Admin])
   @Mutation(() => TranslationModel)
   async deleteTranslation(@Args('id', { type: () => Int }) id: number): Promise<TranslationModel> {
     return this.translationsService.deleteTranslation(id);
